Working Prototype.
This is the initial commit of a scrollable view in a nav bar.
There is a new UX and UI design at Goldstar in which we want to
add multiple sections to the nav bar. A UISegmentedControl won't
work because we're anticipating having more than will fit in an
aesthetically pleasing way. Some of these section names are also
fairly long in relation to what a UISegmentedControl is meant
for.